leetcode/src/main/java/com/wenmq/cn/leetcode/Solution1539.java
package com.wenmq.cn.leetcode;

/**
 * Q1539
 * https://leetcode-cn.com/problems/kth-missing-positive-number/
 */
public class Solution1539 {
    /**
     * 1 <= arr.length <= 1000
     * 1 <= arr[i] <= 1000
     * 1 <= k <= 1000
     * 对于所有 1 <= i < j <= arr.length 的 i 和 j 满足 arr[i] < arr[j] 
     */
    public int findKthPositive(int[] arr, int k) {
        int n = arr.length;
        for (int i = 0; i < n; i++) {
            if (arr[i] - i - 1 >= k) {
                return k + i;
            }
        }
        return k + n;
    }

}
